Add shading and shadows to the scene
I've added the following components to make the scene have proper
lighting and shadows:
* Lights for proper shading
* Invisible light for shadow calculations only
* FilterPostProcessor
  * DirectionalLightShadowFilter
  * BloomFilter
  * SSAOFilter

I changed the logic of extension functions
AssetManager.reflectiveMaterial and AssetManager.unshadedMaterial
to materialize models correctly.

The following instruments were modified to look good given the new
rendering:
* BassDrum only casts shadows (prevents shadows from toms from casting
awkward shadows on to the drum)
* SteamPuffer clouds only cast shadows, preventing awkward shadows
* Fake shadows removed for keyboards, guitars, bass guitars, harps,
mallets, and the drumset
* Space Laser - The laser component and shooter now glow

Also, the skybox needed special shadow settings to allow shadows to
appear in the scene.