Fix shadow issues
Some instruments were not having correct shadows applied because they
were .fbx files? Not sure why this is, so I just converted them to .obj
files and fiddled with the code slightly to work correctly. In the
future, I would like to properly convert them to .j3o binary files.