Improve performance by making sequence length lazy
After performing some profiling, I noticed the call to
Sequencer.microsecondLength was taking a considerable amount of time
per frame, thereby reducing the average FPS. To resolve this, the value
is now calculated lazily--since the length never changes, this should be
safe.

While this is not a direct "end all be all" solution to #115, I can only
imagine that this contributes to the problem presented in that issue,
since it does directly relate to the overall performance of the program.