Remove the "new" display engine
The non-legacy "display engine" used a Swing window to wrap a JmeCanvas
provided by the jMonkeyEngine application. This allowed for better
window handling, but seems to also cause some other issues:

  - When the Windows monitor scale is not set to 100%, the canvas
    incorrectly does not fill the entire JFrame.
  - On some computers, JME will crash when providing a JmeCanvasContext,
    effectively rendering this method useless.

While the conventional method of using the AWT window JME provides has
its own drawbacks:

  - Freezes when the window is being dragged.
  - Window placement is less predictable.

I believe this method is more favorable.

I removed the "legacy" distinction in the settings and the application
will always use this method.

There is a possibility commit may be related to #80 if Swing was not
releasing its resources when the JFrame was destroyed.

Resolves #120