Make some instruments glow (#122)
Add glow effects to the tubular bells and music box when the note is
hit. Also added FifthsKeyboard which glows the key a perfect fourth
down from any playing note (most SoundFonts use a perfect down instead
of a perfect fifth up for the "fifths" instrument, which is odd, but
a P4 down is the same as a P5 up and then an octave down).

Resolves #122