Fix crash when first tempo event is at != 0 time
Fix the issue where an exception is thrown if the first tempo event in
a MIDI file does not occur at time zero.

Per the SMF specification, the tempo shall be assumed to be 120 BPM. The
easiest fix to this is just to add a tempo event of 120 BPM at time zero
if no tempo event occurs then.

Fixes #160