Improve trombone position and animation
Move the trombone to the other side of the stage and closer to the
front, so it can be seen better. Update auto-cam position for trombone.

Separate trombone body and horn so that it stretches like all other
brass instruments.

Improve slide movement anticipation to produce more desirable animation.

Fixes #154