Accurate Keyboard Textures
I've modified the Marimba, Glockenspiel, Xylophone, and Vibraphone textures to look more accurate to actual keyboards and mallets that are used in various percussion environments. I've included two textures for each, the only difference being the mallet. The file names include the model name of the mallet it was inspired by (Most of them are by Vic Firth minus the IP2003), and "light" or "heavy" corresponding to the type of sound they would theoretically produce (Heavy mallets produce louder, sharper sounds while lighter ones are softer)

The multiple textures are mainly because I couldn't decide which mallets should correspond to which instrument, so in the end I guess it's left up to Jacob. If this concept wanted to be expanded upon, you could potentially change which mallets are being used depending on velocity in the MIDI file, but that would obviously require much more effort than just some texture swaps.